coin-back/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlmodel
sqlite-utils
python-dotenv
httpx
python-slugify
alembic